Que no hice aqui

+++ database/migrations/2026_04_08_170718_add_activo_to_users_table.php
@@ -1,6 +1,6 @@
-    <?php
+<?php
 
-    use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
@@ -8,14 +8,18 @@ return new class extends Migration {
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->boolean('activo')->default(1);
+            if (!Schema::hasColumn('users', 'activo')) {
+                $table->boolean('activo')->default(true);
+            }
         });
     }
 
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('activo');
+            if (Schema::hasColumn('users', 'activo')) {
+                $table->dropColumn('activo');
+            }
         });
     }
 };
